actual-requirements-1050.txt
annotated-doc==0.0.4
anyio==4.13.0
audioread==3.1.0
certifi==2022.12.7
cffi==2.0.0
chardet==7.4.0.post2
charset-normalizer==3.4.7
click==8.3.2
contourpy==1.3.3
cycler==0.12.1
cython==3.2.4
decorator==5.2.1
diffusers==0.37.1
einops==0.8.2
fastdtw==0.3.4
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.9.1
idna==3.4
importlib-metadata==7.1.0
importlib-resources==6.5.2
jinja2==3.1.6
joblib==1.5.3
jsonpickle==4.1.1
kiwisolver==1.5.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mido==1.3.3
more-itertools==11.0.1
mpmath==1.3.0
msgpack==1.1.2
music21==9.9.1
networkx==3.6.1
numba==0.65.0
numpy==2.4.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==3.0.2
pillow==12.1.1
platformdirs==4.9.4
pooch==1.9.0
pretty-midi==0.2.11
pyaudio==0.2.14
pycparser==3.0
pygments==2.20.0
pyparsing==3.3.2
pysptk==1.0.1
python-dateutil==2.9.0.post0
pyworld==0.3.5
pyyaml==6.0.3
regex==2026.4.4
requests>=2.31.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.13.1
tgt==1.5
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.66.5
triton==3.1.0
typer==0.24.1
typing-extensions==4.15.0
urllib3==2.6.3
webcolors==25.10.0
zipp==3.19.2
